ComposeMenu: Allow users to pick Multiple Files.
Previously we could only attach single file using the attach file option
in ComposeMenu. This code updates `handleFilesPicker` (previously
`handleFilePicker`) to pick multiple files, its response is an array of
type `DocumentPickerResponse` which is passed to `insertAttachment` to
upload and insert data to `ComposeBox`.

`insertAttachment` is modified to handle this array. We first insert the
`[Uploading ${filename}...]` placeholder for all attachments, then we
upload files one by one and keep replacing their placeholder with
`linkText` on success or `[Uploading ${fileName} failed.]` on failure.

Fixes part of #2366 because behaviour of image picker option is
unchanged.

Fixes-part-of: #2366