replaceRevive tests: Fix bug where "Parse" describe block did nothing.
In my local test runs, at least, I noticed that none of the tests in
the "Parse" describe block were getting run. That's not good! At
least they pass after this commit that fixes that bug, though.

When our code in the "Parse" describe block started running, the
object at `stringified` was empty, so the `.forEach` loop was
iterating through...an empty array.

Our mistake was assuming that the code that built up that object
would be finished running in time. It wasn't finished in time,
because that code was inside a test block of its own. Jest tests
generally run concurrently, so we can't generally expect one test to
be run before another, even if we define the test blocks in a
particular order.