replaceRevive tests: Stop using snapshot tests.
As Greg points out [1], this isn't a best fit for snapshot tests. A
blog post [2] linked from a Jest doc says,

"""
The first thing that became clear to me while using snapshot testing
is that they’re not for everything. They are optimized for a
different case than normal assertion-based tests.

Classic assertion based tests are perfect for testing clearly
defined behavior that is expected to remain relatively stable.

Snapshot tests are great for testing less clearly defined behavior
that may change often.
"""

Which we think is basically right.

[1] https://github.com/zulip/zulip-mobile/pull/4348#discussion_r548762941
[2] https://benmccormick.org/2016/09/19/testing-with-jest-snapshots-first-impressions/