styles [nfc]: Tweak comment on BRAND_COLOR.
Saying what this value is meant to correspond to is important, but the
citation to back up the fact that that's the midpoint can be left to a
commit message.  (The link in the deleted comment works fine, or one
could refer directly to commit 4734533c9.)

A small formatting thing is that when there's a block comment,
anything above the `/*` doesn't look like it's attached to that code.
So a single block comment, or a block of `//` comments, is preferable.