SettingsScreen: Add navigation logic to navigate to 'ThemeScreen'.
Create a new action called 'navigateToThemeScreen' which will help
the 'Theme' OptionButton to navigate to 'ThemeScreen' component
on clicking. Also wired it up with the 'AppNavigator'.

Since, the logic for changing of theme is handled by the 'ThemeScreen' so
there is no use of 'handleThemeChange' function in the 'SettingsScreen'.
Similarly there is no use of Dispatch and the theme prop so all these
are removed in this commit.

Fixes part of #4009