unreads: Don't add messages with "read" flag to unreads.
Previously, the main case where we wouldn't add a message to unreads was
if we were the one who sent it. This check was not correct - instead, we
want to check to see if the message has the "read" flag, which is set by
the server.

In the past, these checks were very close to the same (with the
exception of "Notification Bot" announcing a stream that you created,
which would be "read" for you), but with the addition of muted users,
the check is now incorrect.

This commit removes the logic to not mark a message as read when we are
the one who sent it, and instead just looks at the server flag, since
that should be the canonical place for that information to be.