tools/test [nfc]: Simplify description of default list of suites.
This line was true, but confusing, because no suites are identified
as skipped above.  And they haven't been since the end of the pirlo
suite over a year ago.

I think the reason I didn't simplify that sentence in the first place
when we dropped the pirlo suite was to memorialize the fact that we
*could* have suites that are skipped by default.  A better home for
that information is in the code, as a comment.  So put it there.