RealmInputScreen [nfc]: Remove `realm` route param.
If it seemed weird that the "realm input screen" (the screen that
asks user to put in a realm) was *itself* taking an input called
"realm", that's because it was.

Until this series of commits, `RealmInputScreen` would sometimes
take a valid `realm` on first mount, say "oh, you've got one
already!" and send the user on their way to the auth screen. That
was #4606.

That was fixed when we changed those cases so that they take you
straight to the auth screen without stopping by `RealmInputScreen`.

Now, since there's no use at all to be made of the `realm` param --
now, we're never on this screen with a meaningful value for it --
just remove it. We'll do some further cleanups next.