ios: Modernize some code that's marked as supporting iOS 8 and older.
RN's doc on `Linking` [1] has this note:

> If you also want to listen to incoming app links during your app's
> execution, you'll need to add the following lines to your
> `*AppDelegate.m`:

(We do want to do this; it allows our social auth protocol to work.
So, while we're here, add a comment for that.)

It then gives one snippet marked "iOS 9.x or newer", and another
marked "iOS 8.x and older". We've been using the one for 8.x. So,
switch to the 9.x one.

[1] https://reactnative.dev/docs/0.63/linking