jest config [nfc]: Use a new `projectForPlatform` instead of `common`.
Soon, there'll be a kind of annoying thing we have to do with custom
Jest presets, and we want to have a good place to put a single
long-ish comment about it.

So, make it so that we only have to say `preset: ` once, and plan to
put the comment just above that.