OfflineNotice: Use gray background instead of red, to be less jarring.
An offline state shouldn't be an error / failure state. The app's
capabilities will be reduced, so the user needs to be informed --
but it's not quite right to use a color that conveys that something
unhandled or unexpected has happened.

And switch to the theme-appropriate text color so it shows up in
light mode.