RawLabel [nfc]: Support nesting, for style inheritance.
One bit of `Text`'s API that `RawLabel` hasn't been exposing is its
support for creating an element with `children` and having those
children inherit the element's styles [1]:

> For React Native, we decided to use web paradigm for [displaying
> formatted text] where you can nest text to achieve the same
> effect.

(Before this commit, `RawLabel` was silently accepting `children`
and ignoring it.)

`RawLabel`, unlike `Label`, has the flexibility to match `Text`'s
support for passing some React nodes besides strings. That's because
`RawLabel` is not responsible for passing a string to a function to
get it translated.

While we might have just removed the `text` prop and had all callers
exclusively use `children`, this seems more disruptive than is
worthwhile. Do enforce that exactly one of `children` and `text` is
passed, though.

[1] https://reactnative.dev/docs/text#nested-text