fetchActions tests: Make the `tryFetchFunc`s more parallel.
So it's easier to compare the inputs we're testing. 10ms is a
realistic amount of time for an API request to take. It's also less
than 100ms, which had been used for one of these. So, our tests will
run a bit faster (and they'll run a lot faster when we switch over
to fake timers soon).