fetchActions tests: Test retry logic with a more representative error.
The fact that this test passes isn't good -- it means a basically
arbitrary, unexpected kind of error will let the retry loop continue
without propagating to `tryFetch`'s caller. We'll fix that logic
soon, and add a test case with an error like that. But for now, test
that we get the right behavior with representative inputs.