fetchActions tests: Test retry logic with a more representative error.
The fact that this test (before this commit) was passing isn't good
-- it means a basically arbitrary, unexpected kind of error will let
the retry loop continue without propagating to `tryFetch`'s caller.
Instead, if something unexpected like that happens, we should fail
early (and stop wasting the user's time) by breaking out of the
retry loop and having the error propagate to the caller.

We plan to fix that in a later series of commits, and add a test
case with an error like that. But for now, test that we get the
right behavior with representative inputs.