tests: Convert to eg.mkActionEventNewMessage factory, from fragment.
In particular, drop the occasional `type: EVENT_NEW_MESSAGE` and
`ownUserId: eg.selfUser.user_id`, which were already redundant with
`eg.eventNewMessageActionBase`.

On the other hand, there was one `ownUserId: selfUser.user_id`,
referring to a *different* self-user defined in the specific test.
Happily that now stands out better, making the test clearer.

This should be nearly NFC; the exception is that by using
`eg.mkActionEventNewMessage` we're now making sure `.message.flags`
exists, as it always should on an EVENT_NEW_MESSAGE action.  That
change will in turn be helpful in letting us clean up the code
that these tests are testing.