example data [nfc]: Drop "action fragments" idea, in favor of factories.
As seen in the one existing example of this idea, a factory function
has the advantage that it can handle subtler bits of structure: here,
the fact that the `Message` object found in an `EVENT_NEW_MESSAGE`
action always has a `flags` property, even though other `Message`
objects don't.

As demonstrated in the preceding commit, where we converted uses of
`eg.eventNewMessageActionBase` en masse to `eg.mkActionEventNewMessage`,
it also produces shorter code at the call site.

And the types sure are simpler to describe.