unread: Assert `flags` present on EVENT_NEW_MESSAGE in all subreducers.
This should always be true in real life (as described at the `Message`
definition), and we've just fixed our tests to make it true there too.
So, simplify a bit the potential paths through this code by asserting
the property is there before we try to use it, rather than giving it a
bogus implicit default of all flags being unset.