flags reducer: Assert `flags` present on EVENT_NEW_MESSAGE action.
Much like we did for the various `unread` sub-reducers in the
previous commit, this simplifies a bit the potential paths through
this code.

There was a test for this impossible case; remove that.

While we're here, also remove a similar impossible check on
`messages` being falsy, as if we got an EVENT_UPDATE_MESSAGE_FLAGS
action with no `messages` property.