pm-conversations: Use `useCallback` for an event handler.
This component was converted in a5533a18e from a class component into
a function component.  Before that change, this callback was created
in the class's constructor, so that its value would be constant
through the life of the component.  After that change, it gets a new
value each time the component is rendered, which potentially means
unnecessary re-renders of its children in turn.

The solution to that problem is `useCallback`.  Generally we always
want to use that when translating a constructor-initialized callback
like this onto a function component.  Use it here.