types [nfc]: Unexport OutboxBase
This is really an internal bit of how we go about constructing the
types PmOutbox and StreamOutbox; it doesn't make much sense for types
elsewhere in the application to refer to it.

There was one place we'd been referring to it, at `outboxMessageBase`
in `exampleData`, and that illustrates the point.  Drop that type
annotation entirely.

Like `messagePropertiesBase`, this value is an internal helper used
narrowly just within this module; its real type, in terms of the
promise made in its interface, is "whatever the call site needs".

Flow inherently checks that it satisfies *that* type.  (Try it, after
this change: delete one of the properties in the `outboxMessageBase`
definition, and see an error appear where it's used in `streamOutbox`.
This only works after the preceding commit which fixed
`streamOutbox`'s use of the unsound `$Shape`.)

And on the other hand, writing out what the type works out to
explicitly is kind of verbose and just a recipe for confusion.