auth [nfc]: Move `beginWebAuth` and `endWebAuth` to `webAuth.js`
This change is done to centralise the handling of urls received by
our app which will be required to implement deep linking.

With this modification webAuth can be initiated from any component
with appropriate parameters.

Note that this commit also stops exporting some functions from
`webAuth.js` as this is no longer needed, and updates the jsdocs of
`beginWebAuth` and `endWebAuth`.