Revert "postinstall: Don't run `pod install` unless it's necessary."
This reverts commit 683f3875e.

That quality-of-life change caused a problem, addressed in the
previous commit, where we skipped running `pod install` when we
should have run it. We should have run it because React Native's
autolinking [1] needed to do its job following an update to an NPM
package, by propagating Pod version changes from a .podspec file in
node_modules/ to ios/Podfile.lock.

The conditional wasn't checking for any relevant NPM package updates
since the last time `pod install` was run; it was only checking for
changes to ios/Podfile.lock.

[1] https://github.com/react-native-community/cli/blob/master/docs/autolinking.md