StreamCard [nfc]: Use name/description straight from stream, not subscription.
These properties belong logically to the stream; they're duplicated on
the subscription object but should always have exactly the same value.

In particular, `subscriptionsReducer.js` updates them on a
`stream/update` event just like `streamsReducer.js` does.

So, just use them directly from the Stream object without first
looking for them on the Subscription.

The origin of these fallbacks appears to be c2094ddc7 back in 2017,
where they replaced just looking them up on the Subscription.  That
was indeed buggy, because the Subscription was potentially the bogus
value NULL_SUBSCRIPTION, which has empty name and description.  But
there was never a good reason to look at the Subscription for these
in the first place.