android: Follow RN in changing `versionCodeOverride` calculation.
Done to follow the template-app change in
facebook/react-native@e1bf515ae, on the path to the RN v0.64
upgrade.

The motivation for this change isn't one we share with the template
app: our build process doesn't directly output a different APK per
ABI (CPU architecture) [1], so it doesn't matter if the ABI is
human-readable in the version code. But we might as well follow
along for the sake of staying close to the template app; we don't
see an obvious reason not to.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Multiple.20APKs.3F/near/1131637