MessageList: Stop accepting `typingUsers` prop from parent.
react-redux was injecting a `typingUsers` prop that shadowed this
one, which can cause confusion.

Rather than renaming the outer prop, just remove it.

This gives the same behavior for current callers:
`getCurrentTypingUsers` appropriately contains a sanity check that
will return an empty array when the narrow can't have meaningful
typing state, i.e., when it's not a "conversation" narrow (a PM
narrow or topic narrow) [1]. Search narrows can't have meaningful
typing state, and `MessageList`'s only callsite that passed this
prop was for search narrows.

[1] In fact, it doesn't include topic narrows; we may want to change
    that at some point.