MessageList: Stop accepting `fetching` prop from parent.
react-redux was injecting a `fetching` prop that shadowed this one,
which can cause confusion.

Rather than renaming the outer prop, just remove it.

This gives the same behavior for current callers: the one caller
that passed this prop is for search narrows, and we don't store
fetching state in Redux for search narrows. So we'd go look up the
search narrow in the fetching state, not find it, and default to a
not-fetching value.

This seems better: if, in the future, we do want to start fetching
when you scroll to the top or bottom of a search narrow, we'd want
that to get indicated, just the same way as we do for other narrows.