UserList: Use getItemLayout.
This is a performance improvement, and it also causes the scrollbar to
render correctly (since the UserList knows that all the items have the
same height now).

If you'd like to see this change, look at the scrollbar in the "Group
PM" creation UI.

See the upstream docs on "Optimizing Flatlist Configuration" [1] for
more info about this.

I've chosen to use a library to calculate the correct results for a
SectionList, since that seems a little annoying to write by hand. It's
not been updated in around four years, but I don't see that as a big
problem - I don't expect the way SectionLists work to change in breaking
ways, and if we discover problems, it's a pretty small amount of code
that we could either fork and fix up or rewrite ourselves (around 100
lines).

[1]: https://reactnative.dev/docs/optimizing-flatlist-configuration