tools/test [nfc]: Drop unnecessary quotes.
This is the same syntax we're using in all the `jest_args+=( … )`
lines around this one.  More broadly, inside the parens it works the
same as the syntax one uses for the arguments to any shell command.

Just like in the arguments to any shell command, when the word is
a literal that doesn't need quotes, don't write them.