import: Work around broken imports of '.'.
A bug somewhere in our stack -- presumably Metro or one of its
dependencies -- means that in src/notification/notificationActions.js
the import from `.` was getting resolved not to the module built from
its neighboring `index.js` file, as it should… but to the module
corresponding to the directory `src/common/`.

That also happens to be the directory where our only other imports
from `.` are found.  Seems like the bug may involve the resolution
of `.` being inappropriately shared globally.

In any case, the bug goes away when we make the import reference a
little less special-looking, as `./`.

Thread: https://chat.zulip.org/#narrow/stream/48-mobile/topic/Notification.20Crash/near/1215247
Fixes: #4818