notif [nfc]: Update deprecated code concerning notification token.
The deprecated code was not causing any noticable bugs, but android
studio was prompting that we should fix it, and its good to have up
to date code regardless.

One benefit of this specific commit is that it removes some code.

Something related but not done in this commit is to update
firebase-messaging itself, which as of now stands at 22.0.0 .