notif: Tweak properties of channel.
We change channel name, description and enable lights and vibration
for channel with "default" channel id (which is currently all our
notifications).

Channel name and description are visible in zulip's info in android.
The purpose of channel name is to provide information about the type
of notification so that the user can selectively disable them should
they choose to do so.

Many apps including gmail, slack uses dynamic channel name this
commit does not do that.

A task for future work could be to use different channels for PM and
streams.

reference:
- https://developer.android.com/guide/topics/ui/notifiers/notifications#ManageChannels
- https://developer.android.com/training/notify-user/channels