lightbox [nfc]: Remove meaningless properties in rnfb config.
Done in `downloadFileToCache`, the object passed to
`RNFetchBlob.config` contained some meaningless properties. [1]

These property are technically the sub properties of the property
`addAndroidDownloads` [2]. But we would want to remove them
instead of adding them there since these essentially configure the
android download manager to download the file, which we do not
want in case of this function.

When `addAndroidDownloads` is not set rnfb will use its internal
implementation for downloading based on the `OkHttp` android
library [3], this will not require extra configurations like
setting up the mime type, which if we were to use download
manager, would be required. [4]

Reported-by: Chris Bobbe <cbobbe@zulip.com>

[1] https://chat.zulip.org/#narrow/stream/48-mobile/topic/Sharing.20from.20lightbox.20broken.20on.20Android/near/1223148
[2] https://github.com/joltup/rn-fetch-blob/wiki/Classes#addandroiddownloadsobject-android-only
[3] https://github.com/joltup/rn-fetch-blob/blob/dcbde6f7e12b666b9fe1c8c4a8e2cb04e0048326/android/src/main/java/com/RNFetchBlob/RNFetchBlobReq.java
[4] https://www.npmjs.com/package/rn-fetch-blob/v/0.11.0#android-media-scanner-and-download-manager-support