notif [nfc]: Change key structure of `Recipient.Stream` .
key helps us identify a particular conversation, a conversation is
what we will show as a seperate notification.

Previously all topics in a stream were part of 1 single
conversationa and in this case notification will be unique to a
particular stream, not particular 'stream and topic', which is
really what we should have.