notif: Change the Notification UI to display all conversations.
This commit will make all conversation have their independent
notifications. All messages in a particular conversation will be
seen in a single notification.

It achieves this by using the methods provided by the android API,
specifically `messagingStyle`.

Caveats: There is currently no good way to construct title for
groupPm type conversation. Ideally it should be the names of all the
people in the conversation, but we only have their UserId in the
message payload.