notif: Group Conversations belonging to same realm.
Done by setting the group parameter of notification builder and also
providing a summary notification.

The summary notification is important for the entire notification
functionality to work on all supported devices although it is not
visible in devices with API >= 24. For such devices, android
builds a summary natively.

The summary notification in this commit is just a modified version
of previous single notification we used to show and will show
incorrect information for users with API < 24 and multiple realms.
Hence: Fixes-part-of.

relevant documentation:
- https://developer.android.com/training/notify-user/group

Fixes-part-of: #2691