notif [nfc]: Remove `ByConversationMap`.
`ByConversationMap` was originally intended to store conversations
based on `Identity` as key, so that we would be able to group
conversations by realmId easily.

Here we are removing it, and introducing a new data class in the
following commit that will store realmId and messages along with
notificationId of that particular conversation.

A reason to do this is that structurally we are identifying unique
notification based on its title not by its `Identity`. So its easy
to create a direct mapping of UI notification with the
`ConversationMap` data structure this way. Extra information about
the conversation (eg: realmId) will be included in the value of
ConversationMap (see the next commit).

Skips: #4633