Centerer: Remove, as unnecessary.
AccountPickScreen and AuthScreen both pass `centerContent` to the
`<Screen />` that they use. Empirically, this strategy seems to do
just fine on its own in both cases: removing the `<Centerer />` in
both cases doesn't seem to change anything at all, in my testing.
This is as we'd expect, from reading `Screen`'s interface and noting
that it seems to work just fine without a `<Centerer />` in three
existing uses.

What's more, we have a hypothesis that the `<Centerer />` is causing
an important bug that prevents some users from logging in at all, by
obscuring crucial pieces of the UI; see
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/login.20bug/near/1225893.