createPersistor: Use `storage.multiSet` instead of many `.setItem`s.
As in the previous commit, the hope here is to mitigate #4841.

It doesn't yet solve #4841, though. We can't count on the sets being
done transactionally. Examining `AsyncStorage`'s implementation, we
see that the JavaScript connects to native modules the package
provides, and calls `RNC_AsyncSQLiteDBStorage` and
`RNCAsyncStorage` [1]. These are for Android and iOS, respectively.

The native implementations have these comments for `multiSet`:

Android (from the `ReactContextBaseJavaModule` with name
"RNC_AsyncSQLiteDBStorage" [2]):
```
/**
 * Inserts multiple (key, value) pairs. If one or more of the pairs cannot be inserted, this will
 * return AsyncLocalStorageFailure, but all other pairs will have been inserted.
 * The insertion will replace conflicting (key, value) pairs.
 */
```

iOS (from RNCAsyncStorage.h [3]):
```
// Add multiple key value pairs to the cache.
```

[1] https://github.com/react-native-async-storage/async-storage/blob/v1.6.3/lib/AsyncStorage.js#L17-L18
[2] https://github.com/react-native-async-storage/async-storage/blob/v1.6.3/android/src/main/java/com/reactnativecommunity/asyncstorage/AsyncStorageModule.java#L193-L197
[3] https://github.com/react-native-async-storage/async-storage/blob/v1.6.3/ios/RNCAsyncStorage.h#L40