doEventActionSideEffects: Stop playing in-app "notification" sound.
On iOS, for #4897, we instead want to use the native
notification-sound functionality. So, do that, by adding a line to
AppDelegate.m.

We should still carve out an exception: we won't need to notify the
user if the relevant item (a new or updated message) is already on
the user's screen. Doing so could be annoying; for example, you
don't need to be notified about every single message in an active PM
conversation. That's issue #3114, and it'll take some work. But this
change improves the behavior on iOS and it matches the existing
behavior on Android.

On Android, this JS function call has apparently been duplicating
the notification sound that gets played at a lower level, so it's
fine for Android to remove this.

Fixes: #4897