sharing: Change `SharedFile`.
Done to be better compatible with multiple file sharing (which will
be introduced later on).

Structuring it this way will help us to later on bundle files of all
types inside a single property `files` of `SharedFiles` type where
an individual file type is identified using it's mimeType.

`mimeType` is taken out using contentResolver.getType, in case the
mimeType is unknown it returns null, to handle that we send a
mimetype for denoting binary data (application/octet-stream) when
it does return null.

This makes the code more adaptive since we would not have to create
more types for other files like SharedAudio, SharedImage each time
we try to add custom functionality for such types in the future.

Associated code is changed accordingly to accomodate this change.