api errors: Rename a bit of the Sentry breadcrumb.
The high-order bit about this particular property, relative to
the others in this blob, is that it represents the data the
server sent back to us in reply to the request.  So give it a
name that says that.

A variety of names are possible, but "response" is the term we
use in the types ApiResponse, ApiResponseSuccess, and so on,
which correspond exactly to the data we expect to see here.
So use that.