FullScreenLoading: Stop padding the top inset.
The logo should be centered with respect to the entire screen, not
the safe area (or worse, as it's been doing, the safe area plus
three insets). It's also not big enough to risk entering the insets
on any real screen, so it's fine to ignore the insets here.

Also, this was using `useSafeAreaInsets`, which we try to avoid.