mute [nfc]: Explicitly handle every narrow type in getShownMessagesForNarrow.
Using caseNarrow here takes the case-analysis that was in comments,
and makes it something the type-checker can make sure we haven't
missed any cases of.

In this version we duplicate a bit of code between the stream
and starred cases.  We could factor this differently to avoid that...
but we're about to cut down one of those cases, so don't bother.