mute: Show @-mention messages even if stream or topic muted.
This mirrors the web app's behavior.  For the case of a topic mute,
see `messages_filtered_for_topic_mutes` in `message_list_data.js`.

For the case of a stream mute... the web implementation seems to be
more indirect and I haven't tracked down where stream muting gets
applied to the message list, but empirically it does let messages
through when they have an @-mention.

Fixes: #3472