android: remove default textColor of #0 on Android.
This was introduced in 8919a6aeb, to achieve consistancy with iOS.
With the changes made in 4db472a2e, this becomes the cause of an
inconsistancy where setting system dark mode in android, makes the
text colors black within a dark colored alert dialog box, giving a
poor color contrast and overall bad look to the dialog box.

This commit makes the color white on dark background and black on
light background, based on the theme.

Similar issue on upstream:
https://github.com/facebook/react-native/issues/31345

Discussion on CZO:
https://chat.zulip.org/#narrow/stream/48-mobile/topic/Dark.20theme.20issue.20on.20Android/near/1231579