android: Drop support for API < 23.
As Discussed in CZO [1], Last time (9-2020) we checked, the total
user base using API < 23 is less than 1 percent and hence it is
safe for us to update the API version (according to the update
policy [2] we have).

The advantage of this change would be simplification in some of
the code that we use or will be using, which will be changed or
added in other commits.

The changes in this commit Involved:
* Changing `android/build.gradle`.
* Setting up chrome version to 44 (set according to what is
  written in `platform-versions.md`).
* Modifications to the documentation.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Updating.20.60minSdkVersion.60/near/1236327
[2] https://github.com/zulip/zulip-mobile/blob/master/docs/architecture/platform-versions.md#android-and-ios-versions