online state: Switch to `isInternetReachable`, from `connectionType`.
This is hopefully more accurate about what we actually want to know:
"can the app expect to reach the Internet?" Not just "is there some
network connection?".

Getting this right is already important for a nice user experience.
But it'll be more critical if we adapt our request retry-and-timeout
logic to be aware of the Internet reachability state, which is
reasonable to want to do.

In 6.0.0, `isInternetReachable` started being `boolean | null`; see
the comment about when it's `null`.

Since `null` is a meaningful third state, adjust the Redux plumbing
to allow `null` for `state.session.isOnline` so we can do something
distinct in the UI.