notif: Update Notification UI in android.
This commit will make all conversation have their independent
notifications. All messages in a particular conversation will be
seen in a single notification.

It achieves this by using the methods provided by the android API,
specifically MessagingStyle.

If notification for a given conversation is already present it is
reused (enabling a persistant nature for notification), else a
new notification is created.

Grouping of notification is done using by "groupKey" a combination
of realmUri and userId

Each notification is uniquely identified by a "conversationKey", a
combination of title + realmUri + userId.

Caveats:
* There is currently no good way to construct title for
groupPm type conversation. Ideally it should be the names of all
the people in the conversation, but we only have their UserId in
the message payload.
* This Implementation Removes Dismiss Action Because it will not
  work with the new logic. (functionality wise this shouldn't be
  a concern since users can still dismiss by swiping the
  notification).
* This commit introduces a regression, `RemoveFcmMessage` aren't
  handled appropriately.

Fixes: #2691