async-storage libdef: Align more closely with the doc.
The types that ship with the library are pretty useless; see
  node_modules/@react-native-async-storage/async-storage/lib/commonjs/AsyncStorage.js.flow.

So, check our definitions against the API doc at
  https://react-native-async-storage.github.io/async-storage/docs/api.

That doc isn't very specific either, but at least it confirms that
`getItem` should return a Promise that resolves to `null` or
`string`, and in a roundabout way it confirms that `getAllKeys`
returns a Promise that resolves to an array of `string`s.

For others, it doesn't say what the returned Promise resolves to, so
we say `mixed` for those.