notif [nfc, wip]: Add updateNotification react method and call it.
This is required since reply will perpetually be in loading state
in notification ui unless we update the notification with the new
message.

We create a react method for this, the flow then becomes:
Reply submit starts ReplyService
ReplyService start reply.js
reply.js sends the actual message
reply.js calls this updateNotification to append our message to
the given notification.

Caveats:
* Currently this does not seem to be working
* This still will not work for notification on non active account.