deps: Upgrade react-native-unimodules to 0.14.6, the latest.
And update our setup for Unimodules based on the current doc, at
https://docs.expo.dev/bare/installing-unimodules/. It seems like the
instructions were tweaked upstream since we followed them in
cb87f9062. Might as well update to the new version of the setup
instructions, just with a few intentional deviations:

- The doc asks us to switch some imports of files in node_modules to
  use Node's module resolution algorithm. I confirmed that this new
  approach is recommended (as of a few days ago!), but not
  necessarily required for us if the old way works fine. So, keep
  the old way for now. See
    https://github.com/expo/expo/pull/13901#issuecomment-894530030.

- It asks us to bump `compileSdkVersion` and `targetSdkVersion` from
  29 to 30, in android/build.gradle. I think we generally do this on
  our own schedule.

The only breaking change announced in the changelog (though I should
note that the changelog only goes up to 0.14.1) is that the exported
`Permissions` object is removed, and `expo-permissions` is removed
from dependencies. That package has been retired, and its
functionality has moved to the relevant Expo packages that request
certain permissions (like `expo-location`, `expo-camera`, etc.). See
  https://github.com/expo/expo/blob/master/packages/react-native-unimodules/CHANGELOG.md.
That's all fine; we haven't been using any of that.