deps: Remove expo-splash-screen.
This seems to be no longer needed: we added it in c4fca9d9d to
satisfy a peer-dependency constraint, but now I'm not getting a
peer-dep warning if I remove it.

See #4279 and discussion at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/React.20Navigation.20v5/near/1042787
for some warnings from Greg about the design of expo-splash-screen,
should we choose to actually use it one day.