deps: Upgrade color to 4.0.1, the latest.
They make the following announcement for major release 4.0.0 [1]:

> NOTE: "Major" here used loosely. This release is an interim major
> release that introduces ES6 syntax into the package without
> changing the import method (require()). There will be a follow-up
> major release that switches entirely to ESM and will set the
> appropriate engine key in package.json. This was just a necessary
> first step to allow those who don't (yet) use ESModules a sane
> place to upgrade and pin to.

We're fine with ES6 syntax. So we shouldn't have to do anything
special to take this upgrade.

...Ah, but we're not yet fine with some *ES12* syntax that was snuck
in unannounced: the "NumericLiteralSeparator" [2] [3]. So, bring in
a Babel plugin for that, for now. I see this has caused some
controversy:
  - https://github.com/Qix-/color/issues/204
  - https://github.com/Qix-/color/issues/206
  - https://github.com/Qix-/color/pull/208
  - https://github.com/Qix-/color/issues/209

I seemed to need to run `yarn start` (or `react-native start`) with
`--reset-cache` before the issue went away when testing on Android
outside "Debug" / "Debug JS Remotely" mode.

The error was the following:

  [SyntaxError: No identifiers allowed directly after numeric
  literal]

Ah, well, easy enough to bring in the plugin and use
`--reset-cache`, even if the debugging path to get there was kind of
long.

[1] https://github.com/Qix-/color/releases/tag/4.0.0
[2] https://262.ecma-international.org/12.0/#prod-NumericLiteralSeparator
[3] https://github.com/Qix-/color/blob/4.0.1/index.js#L257