deps: Upgrade react-native-reanimated to 2.2.0, the latest.
The release notes [1] don't say "breaking" for any of the newer
versions, so I guess this is safe to do? We also don't pull in
`react-native-reanimated` to use it in any of our own code.

It's a peer dependency of some React Navigation libraries; this is
what I see if I remove it:

warning " > @react-navigation/drawer@5.12.5" has unmet peer
  dependency "react-native-reanimated@>= 1.0.0".
warning " > @react-navigation/material-top-tabs@5.3.15" has unmet
  peer dependency "react-native-reanimated@>= 1.0.0".
warning " > react-native-tab-view@2.16.0" has unmet peer dependency
  "react-native-reanimated@*".

So, those libraries have the risky, fanciful practice [2] of
declaring themselves compatible with all future versions of
react-native-reanimated, including this one…

…and then patching incompatibilities when they are predictably
reported:

- @react-navigation/drawer@ea5affd91, released in
  @react-navigation/drawer@5.8.3 (we're on a version later than
  that)

- satya164/react-native-tab-view@0996cdf48, released in
  satya164/react-native-tab-view@2.14.3 (we're on a version later
  than that)

Still, react-native-reanimated v2.0.0 was released in 2021-03-04,
which is already several months old. So, while we don't find a
credible declaration of how likely things are to break, at least
these packages have had some time to fix any breakages that were
reported, and we'd have taken those fixes (or we will soon) because
we're now in the habit of running `yarn upgrade` regularly.

[1] https://github.com/software-mansion/react-native-reanimated/releases
[2] https://github.com/react-native-webview/react-native-webview/pull/2054#issuecomment-895435000