deps: Remove string.fromcodepoint.
For future polyfills we might need, let's try following the
(deprecated) package @babel/polyfill's advice and use `core-js`; see
  https://babeljs.io/docs/en/babel-polyfill.

This is an old ES6 feature that's been out for a long time. MDN
says [1] that "Safari on iOS" gained support in version 10, which we
take to mean our app would work on iOS 10+ because we use the
system's JS engine, which is presumably the same as the JS engine
used in Safari. We support iOS 12.1+, and this commit worked in a
simulator running iOS 12.4, so this is fine.

We follow a kind of annoying process for confirming support on
Android [2] by reading WebKit changelogs. We also double-check with
an experiment: if the code works without the polyfill at this point,
on any Android device, then it should be fine for all Android users
because they all get the same JavaScriptCore that ships with the
app. And it did work; I tested on the office Android device.

In the experiments on the iOS 12.4 simulator and the office Android
device, I made sure to turn "Debug" (formerly "Debug JS Remotely")
off. If I hadn't, then the JavaScript would have been running in
Chrome on my computer, not on the device, defeating the experiment.

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/fromCodePoint#browser_compatibility
[2] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Can.20remove.20.60string.2Efromcodepoint.60.3F/near/1246051