i18n [nfc]: Say "selfname" for a language's name for itself.
I've never been real thrilled with calling this a "native name",
which doesn't sound right.  One square-on term for it would be
"endonym", or "autonym", but those are a bit fancy.

Then a few weeks ago we were discussing how to name a similar
property in the Zulip API, and I discovered that there is an accepted
term with the same meaning which is rather more transparent!
  https://chat.zulip.org/#narrow/stream/378-api-design/topic/language.20parameters.20in.20page_params/near/1211764
  https://en.wiktionary.org/wiki/selfname

That Wiktionary entry quotes the term being used in two different
books in the professional linguistic literature, and by different
authors; so it seems it's just as respectable as any fancy
Greek-derived term, while being a lot more transparent than those.