user: Check for an active account in getHaveServerData.
This isn't itself server data -- it's what we use for acquiring
server data -- which is why we didn't cover it when expanding these
checks in 916dc4b28.

But like the server data, it's something we assume is present
when rendering the main UI of the app.  And because of the same
issue #4841, the same lack of consistency in our data-storage
model with redux-persist that caused #4587 and made that fix
necessary, it's possible for it to be lacking data when the
existing checks in this function all pass.

So, add a check for this too.  Build on the comments added in the
previous commit to reason through why this won't get us stuck at
the loading screen; and add a comment at the otherwise
inessential-looking bit of code that provides one of the
assumptions we rely on for that reasoning.

Also correct a related point in the comment at the end of the
function.