ServerCompatBanner: Bump `minSupportedVersion` to 2.1.0, as planned.
We haven't gotten negative feedback about the banner from the
smaller pool of users connecting to servers earlier than 2.0.0. The
banner was introduced with that threshold (in PR #4750) to beta
users in v27.164 (2021-06-17) and to all users in v27.165
(2021-06-24).

So, proceed with aligning our threshold with what we declare in the
server stable-release doc, and update the code comment to match.