initialFetchAbort: Don't kick user out if we have server data.
I have a draft in progress for a better solution to this; see
discussion at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/connection.20failed/near/1248170.

But for now this should fix an annoying situation that's so far been
observed on Alya's Android device, where a "Gave up trying to
connect" alert appears immediately on startup. The theory is that
the app stops making progress on the network request when the app is
backgrounded, while of course it doesn't stop approaching the
timeout. Then, when the app is foregrounded again, the request
hasn't succeeded but the timeout has fired, and the alert is shown.

Tested manually by using a server that's become unavailable (and
reducing the timeout in the code to 10 seconds to speed up the test).
We keep showing the main app UI as desired, and the debug console
shows the intended sequence of DEAD_QUEUE and new INITIAL_FETCH_START
actions happening.