ios notif: Require realm_uri in payload, present since server 1.9.
This release has been out for nearly three years (since 2018-11-07),
so it's long past time for anyone running an older version to upgrade.
We decided earlier this year to stop supporting older than 2.1.x; and
since the mobile release v27.165 of a couple of months ago, anyone
using a server older than 2.0.x has gotten a warning banner about it.

As of this commit, our post-parse `Notification` type still sees
`realm_uri` as optional.  We'll tighten that after making it similarly
required in the Android-native parsing code.